+++ ce_ui/views.py
@@ -1,4 +1,6 @@
+import csv
 import logging
+import os
 from html import unescape
 
 from allauth.account.views import EmailView
@@ -23,6 +25,8 @@ from topobank_publication.models import PublicationCollection
 from topobank_publication.serializers import PublicationCollectionSerializer
 from topobank.manager.v1.serializers import SurfaceSerializer, TopographySerializer
 from topobank.manager.utils import subjects_from_base64, subjects_to_base64
+from topobank.manager.v1.serializers import (SurfaceSerializer,
+                                             TopographySerializer)
 from topobank.usage_stats.utils import increase_statistics_by_date
 from topobank.users.models import User
 from trackstats.models import Metric, Period
@@ -295,7 +299,7 @@ def extra_tabs_if_single_item_selected(context, subjects):
 
 
 class AnalysisDetailView(AppDetailView):
-    model = AnalysisFunction
+    model = Workflow
     slug_field = "name"
     vue_component = "AnalysisDetail"
     serializer_class = WorkflowSerializer
@@ -563,3 +567,26 @@ class UserListView(LoginRequiredMixin, ListView):
     # These next two lines tell the view to index lookups by username
     slug_field = "username"
     slug_url_kwarg = "username"
+
+
+class ChallengeHomepageView(TemplateView):
+    template_name = "challenge/homepage.html"
+
+
+class ChallengeListOfPublishedDataView(TemplateView):
+    template_name = "challenge/list_of_published_data.html"
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+
+        data_list = {}
+        csv_path = os.path.join(os.path.dirname(__file__), "data", "challenge_data.csv")
+
+        with open(csv_path, mode="r") as file:
+            reader = csv.DictReader(file)
+
+            for row in reader:
+                sample_id = row["sample_id"]
+                data_list[sample_id] = {"doi_link": row["doi_link"]}
+        context["data_list"] = data_list
+        return context